M2-5204: Move answers fixtures to the conftest
